move direct upload step def. into own context
