+++ tests/acceptance/features/bootstrap/FeatureContext.php
@@ -26,8 +26,6 @@ class FeatureContext implements Context {
 
 	private ?ResponseInterface $response = null;
 
-	private string $lastCreatedDirectUploadToken = '';
-
 	public function getAdminUsername(): string {
 		return $this->adminUsername;
 	}
@@ -388,7 +386,7 @@ class FeatureContext implements Context {
 		return $this->getIdOfElement($user, $destination);
 	}
 
-	private function getIdOfElement(string $user, string $element): int {
+	public function getIdOfElement(string $user, string $element): int {
 		$propfindResponse = $this->makeDavRequest(
 			$user,
 			$this->regularUserPassword,
@@ -616,57 +614,6 @@ class FeatureContext implements Context {
 	}
 
 	/**
-	 * @When /^user "([^"]*)" sends a GET request to the direct\-upload endpoint with the ID of "(.*)"$/
-	 */
-	public function userSendsAGETRequestToTheEndpointWithTheFileIdOf(
-		string $user, string $elementName
-	): void {
-		$elementId = $this->getIdOfElement($user, $elementName);
-		$this->sendRequestsToAppEndpoint(
-			$user,
-			$this->regularUserPassword,
-			'GET',
-			'direct-upload?folder_id=' . $elementId
-		);
-	}
-
-	/**
-	 * @When /^user "([^"]*)" sends a GET request to the direct\-upload endpoint with the ID "(.*)"$/
-	 */
-	public function userSendsAGETRequestToTheEndpointWithTheId(
-		string $user, string $folderId
-	): void {
-		$this->sendRequestsToAppEndpoint(
-			$user,
-			$this->regularUserPassword,
-			'GET',
-			'direct-upload?folder_id=' . $folderId
-		);
-	}
-
-	/**
-	 * @Given /^user "([^"]*)" got a direct-upload token for "(.*)"$/
-	 */
-	public function userGotADirectUploadTokenFor(
-		string $user, string $elementName
-	): void {
-		$elementId = $this->getIdOfElement($user, $elementName);
-		$this->sendRequestsToAppEndpoint(
-			$user,
-			$this->regularUserPassword,
-			'GET',
-			'direct-upload?folder_id=' . $elementId
-		);
-		$this->theHttpStatusCodeShouldBe(200);
-		$responseAsJson = json_decode($this->response->getBody()->getContents());
-		Assert::assertObjectHasAttribute(
-			'token', $responseAsJson,
-			'cannot find token in response'
-		);
-		$this->lastCreatedDirectUploadToken = $responseAsJson->token;
-	}
-
-	/**
 	 * @param string $username
 	 * @param string $password
 	 * @param string $method
@@ -675,7 +622,7 @@ class FeatureContext implements Context {
 	 * @return void
 	 * @throws \GuzzleHttp\Exception\GuzzleException
 	 */
-	private function sendRequestsToAppEndpoint(
+	public function sendRequestsToAppEndpoint(
 		string $username,
 		string $password,
 		string $method,
